+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -5,6 +5,11 @@
 - Breaking: `ngxSplit` directives no longer require, nor support, @angular/flex-layout directives
 - Enhancement: Remove @angular/flex-layout
 - Enhancement: Added `ngx-flex` classes for easy flex layouts
+
+## 28.6.0 (2020-05-07)
+
+- Feature: New ngx-navbar component (#437)
+- Fix: Update tab buttons when disabled input changes on ngx-tab (#438, fixes #436)
 - Fix: allow null values from numeric ngx-inputs (#440)
 - Fix: ngx-select Set touched value inside onTouchedCallback (#441)